+++ packages/next-common/components/statistics/democracy/beenDelegatedPopup/delegationTabList.js
@@ -1,10 +1,9 @@
-import React, { useEffect, useState } from "react";
+import { useEffect, useState } from "react";
 import Pagination from "next-common/components/pagination";
 import VStack from "next-common/components/styled/vStack";
 import styled from "styled-components";
 import nextApi from "next-common/services/nextApi";
 import { EmptyList } from "next-common/utils/constants";
-import StyledList from "next-common/components/styledList";
 import useColumns from "next-common/components/styledList/useColumns";
 import Flex from "next-common/components/styled/flex";
 import ValueDisplay from "next-common/components/valueDisplay";
@@ -13,6 +12,7 @@ import { useChainSettings } from "next-common/context/chain";
 import { convictionToLockX } from "next-common/utils/referendumCommon";
 import PopupListWrapper from "next-common/components/styled/popupListWrapper";
 import AddressUser from "next-common/components/user/addressUser";
+import DataList from "next-common/components/dataList";
 
 const ConvictionText = styled.span`
   width: 40px;
@@ -127,7 +127,7 @@ export default function DelegationTabList({
   return (
     <VStack space={16}>
       <MyPopupListWrapper>
-        <StyledList columns={columns} rows={rows} loading={!isLoaded} />
+        <DataList columns={columns} rows={rows} loading={!isLoaded} />
       </MyPopupListWrapper>
       <Pagination {...pagination} />
     </VStack>
