+++ packages/next-common/components/summary/democracyBeenDelegated/beenDelegatedListPopup/delegationList.js
@@ -1,12 +1,11 @@
-import React from "react";
 import { toPrecision } from "next-common/utils";
 import ValueDisplay from "next-common/components/valueDisplay";
 
 import { useChainSettings } from "next-common/context/chain";
 import PopupListWrapper from "../../../styled/popupListWrapper";
-import StyledList from "next-common/components/styledList";
 import CapitalTableItem from "next-common/components/popup/capitalTableItem";
 import AddressUser from "next-common/components/user/addressUser";
+import DataList from "next-common/components/dataList";
 
 function DelegationList({ items, loading = false }) {
   const node = useChainSettings();
@@ -69,7 +68,7 @@ function DelegationList({ items, loading = false }) {
 
   return (
     <PopupListWrapper>
-      <StyledList columns={columns} rows={rows} loading={loading} />
+      <DataList columns={columns} rows={rows} loading={loading} />
     </PopupListWrapper>
   );
 }
